feat: made the project a maven project
