+++ .idea/libraries/Maven__org_eclipse_jetty_jetty_server_7_0_2_v20100331.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: org.eclipse.jetty:jetty-server:7.0.2.v20100331">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/eclipse/jetty/jetty-server/7.0.2.v20100331/jetty-server-7.0.2.v20100331.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/org/eclipse/jetty/jetty-server/7.0.2.v20100331/jetty-server-7.0.2.v20100331-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/eclipse/jetty/jetty-server/7.0.2.v20100331/jetty-server-7.0.2.v20100331-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
