+++ .idea/libraries/Maven__org_mortbay_jetty_servlet_api_3_0_PFD20090525.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: org.mortbay.jetty:servlet-api:3.0.PFD20090525">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/mortbay/jetty/servlet-api/3.0.PFD20090525/servlet-api-3.0.PFD20090525.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/org/mortbay/jetty/servlet-api/3.0.PFD20090525/servlet-api-3.0.PFD20090525-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/mortbay/jetty/servlet-api/3.0.PFD20090525/servlet-api-3.0.PFD20090525-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
