+++ src/main/java/Continuous_Integration1.iml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module org.jetbrains.idea.maven.project.MavenProjectsManager.isMavenModule="true" type="JAVA_MODULE" version="4">
+  <component name="NewModuleRootManager" LANGUAGE_LEVEL="JDK_1_8">
+    <output url="file://$MODULE_DIR$/target/classes" />
+    <output-test url="file://$MODULE_DIR$/target/test-classes" />
+    <content url="file://$MODULE_DIR$">
+      <sourceFolder url="file://$MODULE_DIR$/src/main/java" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/src/main/resources" type="java-resource" />
+      <sourceFolder url="file://$MODULE_DIR$/src/test/java" isTestSource="true" />
+      <excludeFolder url="file://$MODULE_DIR$/target" />
+    </content>
+    <orderEntry type="inheritedJdk" />
+    <orderEntry type="sourceFolder" forTests="false" />
+    <orderEntry type="library" name="Continuous_Integration1" level="project" />
+    <orderEntry type="library" name="Maven: org.eclipse.jetty:jetty-server:7.0.2.v20100331" level="project" />
+    <orderEntry type="library" name="Maven: javax.servlet:servlet-api:2.5" level="project" />
+    <orderEntry type="library" name="Maven: org.eclipse.jetty:jetty-continuation:7.0.2.v20100331" level="project" />
+    <orderEntry type="library" name="Maven: org.eclipse.jetty:jetty-http:7.0.2.v20100331" level="project" />
+    <orderEntry type="library" name="Maven: org.eclipse.jetty:jetty-io:7.0.2.v20100331" level="project" />
+    <orderEntry type="library" name="Maven: org.eclipse.jetty:jetty-util:7.0.2.v20100331" level="project" />
+    <orderEntry type="library" name="Maven: org.mortbay.jetty:servlet-api:3.0.PFD20090525" level="project" />
+    <orderEntry type="library" scope="TEST" name="Maven: junit:junit:4.13.2" level="project" />
+    <orderEntry type="library" scope="TEST" name="Maven: org.hamcrest:hamcrest-core:1.3" level="project" />
+  </component>
+</module>
